feat(php-ini): edit the version and shared ini side by side
The php.ini tab edited one file at a time behind a scope dropdown. The
two files work together, the shared file sets the baseline and the
version file overrides it, so reading one while editing the other meant
switching scope and reloading the editor, which also dropped whatever
was in the buffer.

The tab now splits vertically: the version's own ini on the left, the
shared ini on the right, stacking below lg. The per-file logic moved
into PhpIniPane, so each pane owns its own load, dirty state, save,
revert, reset and backup restore, and an action on one pane cannot
touch unsaved edits in the other.

Both panes read the scope labels the dropdown used to show, so only the
now-unused "Editing" string is dropped from the locale files.
